🚀 App-Vigilante: Migrações e Seeders Completos
✅ **Migrações Criadas:**
- usuario: Campos completos com tipos e autenticação
- posto_trabalho: Postos de vigilância
- ponto_base: Pontos de verificação com coordenadas
- cartao_programas: Programas de trabalho
- cartao_programa_pontos: Ciclos de pontos com horários
- escala: Escalas dos vigilantes
- escala_diaria: Controle diário das escalas
- registro_presenca: Registros de chegada/saída

✅ **Seeders Implementados:**
- UsuarioSeeder: Vigilantes e admin
- PostoTrabalhoSeeder: Postos padrão
- PontoBaseSeeder: Pontos com coordenadas
- Relacionamentos e foreign keys corretos

✅ **Correções de Models:**
- Campo posto_trabalho_id corrigido em PontoBase
- Relacionamentos atualizados
- Campos fillable completos

✅ **Funcionalidades:**
- 10 ciclos de pontos com horários específicos
- Sistema de presença funcionando
- Dashboard com feed da comunidade
- Erro 419 CSRF resolvido
- Card horário removido, nome vigilante simplificado

🎯 **Banco Estruturado:**
- Sequência correta de migrações
- Dados de teste consistentes
- Foreign keys e índices otimizados

+++ app-morador/database/migrations/2025_08_28_164657_create_alertas_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->enum('tipo', ['seguranca', 'manutencao', 'geral', 'emergencia']);
             $table->enum('prioridade', ['baixa', 'media', 'alta', 'critica']);
             $table->enum('status', ['ativo', 'resolvido', 'cancelado']);
-            $table->unsignedInteger('usuario_id')->nullable(); // ID do vigilante que criou
+            $table->unsignedBigInteger('usuario_id')->nullable(); // ID do vigilante que criou
             $table->string('localizacao')->nullable();
             $table->json('coordenadas')->nullable(); // latitude e longitude
             $table->timestamp('resolvido_em')->nullable();
